+++ normalize/lib/normalize.dart
@@ -4,8 +4,8 @@ export 'src/denormalize_operation.dart';
 export 'src/denormalize_fragment.dart';
 export 'src/normalize_fragment.dart';
 
-export 'src/options/field_policy.dart';
-export 'src/options/type_policy.dart';
+export 'src/policies/field_policy.dart';
+export 'src/policies/type_policy.dart';
 export 'src/utils/add_typename_visitor.dart';
 export 'src/utils/exceptions.dart';
 export 'src/utils/get_operation_definition.dart';