+++ normalize/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2020-09-08
+
+### Added
+
+- add `FieldKey` class
+
 ## [0.3.0] - 2020-09-05
 
 ### Changed